lex: Remove quotes from quoted idenfitiers in the parser
Fixes https://github.com/saresend/dotfile-parser/issues/8

It is not done in the lexer, as the &str is copied around, while
copying String is more expensive. In the parser it's the last step
before consuming user code.

It is a bit fragile though, to remember to call unquote_string instead
of .to_owned() or .to_string().